+++ R/add_labels.R
@@ -65,10 +65,10 @@ add_labels <- function(x, value) {
   if (is.null(value)) stop("`value` is NULL.", call. = F)
   if (is.null(names(value))) stop("`value` must be a named vector.", call. = F)
 
-  if (is.matrix(x) || is.data.frame(x) || is.list(x)) {
+  if (is.data.frame(x) || is.list(x)) {
     # get length of data frame or list, i.e.
     # determine number of variables
-    if (is.data.frame(x) || is.matrix(x))
+    if (is.data.frame(x))
       nvars <- ncol(x)
     else
       nvars <- length(x)