+++ R/dicho.R
@@ -70,10 +70,10 @@ dicho <- function(x,
     stop("argument `dich.by` must either be `median`, `mean` or a numerical value." , call. = FALSE)
   }
 
-  if (is.matrix(x) || is.data.frame(x) || is.list(x)) {
+  if (is.data.frame(x) || is.list(x)) {
     # get length of data frame or list, i.e.
     # determine number of variables
-    if (is.data.frame(x) || is.matrix(x))
+    if (is.data.frame(x))
       nvars <- ncol(x)
     else
       nvars <- length(x)