+++ R/replace_na.R
@@ -78,10 +78,10 @@
 replace_na <- function(x, value, na.label = NULL, tagged.na = NULL) {
   # check for valid value
   if (is.null(value) || is.na(value)) return(x)
-  if (is.matrix(x) || is.data.frame(x) || is.list(x)) {
+  if (is.data.frame(x) || is.list(x)) {
     # get length of data frame or list, i.e.
     # determine number of variables
-    if (is.data.frame(x) || is.matrix(x))
+    if (is.data.frame(x))
       nvars <- ncol(x)
     else
       nvars <- length(x)