+++ NEWS.md
@@ -12,3 +12,4 @@
 ## Bug fixes
 
 * `is_empty` returned `NA` instead of `TRUE` for empty character vectors.
+* Fixed bug with erroneous assignment of value labels to subset data when using `copy_labels` ([#20](https://github.com/sjPlot/sjmisc/issues/20))