+++ NEWS.md
@@ -2,9 +2,13 @@
 
 ## General
 
-* All statistical functions have been removed and are now in a separate package, `sjstats` (https://cran.r-project.org/package=sjstats).
+* All statistical functions have been removed and are now in a separate package, [sjstats](https://cran.r-project.org/package=sjstats).
 * Removed some S3-methods for `labelled`-class, as these are now provided by the haven-package.
 * **sjmisc** now supports _tagged_ `NA` values, a new structue for missing values introduced by the [haven-package](https://cran.r-project.org/package=haven). This means that functions or arguments that are no longer useful, have been removed while other functions dealing with NA values have been largely revised.
+
+## Changes to functions
+
+* `str_contains` gets a `switch` argument to switch the role of `x` and `pattern`.
 
 ## New functions