+++ NEWS.md
@@ -2,9 +2,10 @@
 
 ## General
 
-* **sjmisc** now supports _tagged_ `NA` values, a new structue for labelled missing values introduced by the [haven-package](https://cran.r-project.org/package=haven). This means that functions or arguments that are no longer useful, have been removed while other functions dealing with NA values have been largely revised.
+* **sjmisc** now supports _tagged_ `NA` values, a new structue for labelled missing values introduced by the [haven-package](https://cran.r-project.org/package=haven). This means that functions or arguments that are no longer useful, have been removed while other functions dealing with NA values have been largely revised:
 * All statistical functions have been removed and are now in a separate package, [sjstats](https://cran.r-project.org/package=sjstats).
 * Removed some S3-methods for `labelled`-class, as these are now provided by the haven-package.
+* Functions no longer check input for type `matrix`, to avoid conflicts with scaled vectors (that were recognized as matrix and hence treated as data frame).
 
 ## New functions